task 6

+++ main.h
@@ -26,7 +26,7 @@ int print_pointer(va_list val);
 int print_unsigned(va_list args);
 int print_hex_extra(unsigned long int num);
 int _strlenc(const char *s);
-int print_HEX_extra(unsigned int num);
+int print_HEX_extra(unsigned long int num);
 int print_exc_string(va_list val);
 int print_HEX(va_list val);
 int print_hex(va_list val);
